+++ hunch_app/lib/chat_bubble.dart
@@ -1,21 +1,21 @@
-import 'package:flutter/material.dart';
+// import 'package:flutter/material.dart';
 
-class ChatBubble extends StatelessWidget {
-  final String message;
-  const ChatBubble({super.key, required this.message});
+// class ChatBubble extends StatelessWidget {
+//   final String message;
+//   const ChatBubble({super.key, required this.message});
 
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(12),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(8),
-        color: Colors.blue,
-      ),
-      child: Text(
-        message,
-        style:  const TextStyle(fontSize: 16, color: Colors.white),
-      ),
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return Container(
+//       padding: const EdgeInsets.all(12),
+//       decoration: BoxDecoration(
+//         borderRadius: BorderRadius.circular(8),
+//         color: Colors.blue,
+//       ),
+//       child: Text(
+//         message,
+//         style:  const TextStyle(fontSize: 16, color: Colors.white),
+//       ),
+//     );
+//   }
+// }
