+++ hunch_app/lib/screens/Home_screen.dart
@@ -2,10 +2,12 @@ import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:hunch_app/Chatpage.dart';
+import 'package:hunch_app/chat/Chatpage.dart';
 import 'package:hunch_app/screens/LoginPage.dart';
 import 'package:hunch_app/auth_service.dart';
-import 'package:hunch_app/chat_page.dart';
+import 'package:hunch_app/chat/chat_page.dart';
+import 'package:hunch_app/users.dart/userPage.dart';
+import 'package:hunch_app/users.dart/user_service.dart';
 import 'package:provider/provider.dart';
 
 // Define the MenuAction enum here
@@ -14,34 +16,28 @@ enum MenuAction {
   // Add more menu actions if needed
 }
 
-
 class HomeScreen extends StatefulWidget {
-  const HomeScreen({Key ? key}): super(key: key);
+  const HomeScreen({Key? key}) : super(key: key);
 
   @override
   State<HomeScreen> createState() => _HomeScreenState();
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-
-   int _currentIndex = 0;
+  int _currentIndex = 0;
 
   final List<Widget> _screens = [
     Container(
-      child: CircularProgressIndicator(),
+      child: const CircularProgressIndicator(),
     ),
     Container(
       color: Colors.amber,
     ),
     Container(),
+
     Chatpage(),
-    Container(),
-
-
-    
+    UserPage()
   ];
-
-
 
   @override
   Widget build(BuildContext context) {
@@ -49,32 +45,25 @@ class _HomeScreenState extends State<HomeScreen> {
         onWillPop: () async {
           return await showExitConfirmationDialog(context);
         },
-    
-    
-   child: Container(
-      decoration: const BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topCenter,
-          end: Alignment.bottomCenter,
-          stops: [0.4, 0.7],
-          colors: <Color>[
-                    Colors.white,
-                    Color.fromARGB(255, 148, 213, 237),
-                    
-
-                  ]
-
-        )
-      ),
-      child: Scaffold(
-        backgroundColor: Colors.transparent,
-        
-        appBar: AppBar(
-
-           actions: [
+        child: Container(
+          decoration: const BoxDecoration(
+              gradient: LinearGradient(
+                  begin: Alignment.topCenter,
+                  end: Alignment.bottomCenter,
+                  stops: [
+                0.4,
+                0.7
+              ],
+                  colors: <Color>[
+                Colors.white,
+                Color.fromARGB(255, 255, 255, 255),
+              ])),
+          child: Scaffold(
+            backgroundColor: Colors.transparent,
+            appBar: AppBar(
+              actions: [
                 PopupMenuButton<MenuAction>(
                   onSelected: (value) async {
-                    
                     switch (value) {
                       case MenuAction.signUserOut:
                         final shouldLogout = await showSignOutDialog(context);
@@ -95,38 +84,29 @@ class _HomeScreenState extends State<HomeScreen> {
                   },
                 ),
               ],
-          title: RichText(
-            text: 
-              TextSpan(
-                text: "H",
-                style: GoogleFonts.ubuntu(
-                  textStyle: const TextStyle(
-                    color: Color.fromARGB(255, 224, 2, 2),
-                    fontSize: 24,
-                    fontWeight: FontWeight.bold,
-    
-                  ),
-                ),
-                children: <TextSpan>[
-                  TextSpan(
-                    text: "unch App",
-                    style: GoogleFonts.ubuntu(
-                      textStyle: TextStyle(
-                        fontSize: 24,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.black
+              title: RichText(
+                  text: TextSpan(
+                      text: "H",
+                      style: GoogleFonts.ubuntu(
+                        textStyle: const TextStyle(
+                          color: Color.fromARGB(255, 224, 2, 2),
+                          fontSize: 24,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      children: <TextSpan>[
+                    TextSpan(
+                      text: "unch App",
+                      style: GoogleFonts.ubuntu(
+                        textStyle: TextStyle(
+                            fontSize: 24,
+                            fontWeight: FontWeight.bold,
+                            color: Colors.black),
                       ),
                     ),
-                  ),
-    
-                ]
-              )
-            
-        ),
-
-        
-      ),
-      drawer: Drawer(
+                  ])),
+            ),
+            drawer: Drawer(
               child: ListView(
                 padding: EdgeInsets.zero,
                 children: const <Widget>[
@@ -136,8 +116,7 @@ class _HomeScreenState extends State<HomeScreen> {
                         accountName: Text("Gaurav"),
                         accountEmail: Text("gaurav2211028@akgec.ac.in"),
                         currentAccountPicture: CircleAvatar(
-                          backgroundImage:
-                              AssetImage('assets/images/home.jpg'),
+                          backgroundImage: AssetImage('assets/images/home.jpg'),
                         ),
                       )),
                   ListTile(
@@ -167,43 +146,42 @@ class _HomeScreenState extends State<HomeScreen> {
                 ],
               ),
             ),
-
-            body:_screens[_currentIndex],
-
-      bottomNavigationBar: CurvedNavigationBar(
-        backgroundColor: Colors.transparent,
-        color: const Color.fromARGB(255, 169, 146, 233),
-        animationDuration: const Duration(milliseconds: 400),
-        onTap: (index){
-          setState(() {
-              _currentIndex = index;
-            });
-
-        },
-        items: const [
-          Icon(Icons.home,
-          color: Color.fromARGB(255, 255, 255, 255),),
-          Icon(Icons.search,
-          color: Color.fromARGB(255, 255, 255, 255),),
-          Icon(Icons.add,
-          color: Color.fromARGB(255, 255, 255, 255),),
-          Icon(Icons.message,
-                 color: Color.fromRGBO(255, 255, 255, 1),),
-          Icon(Icons.person,
-          color: Color.fromARGB(255, 255, 255, 255),),
-    
-    
-        ]
-      ) ,
-      ),
-   )
-    );
+            body: _screens[_currentIndex],
+            bottomNavigationBar: CurvedNavigationBar(
+                backgroundColor: Colors.transparent,
+                color: const Color.fromARGB(255, 169, 146, 233),
+                animationDuration: const Duration(milliseconds: 400),
+                onTap: (index) {
+                  setState(() {
+                    _currentIndex = index;
+                  });
+                },
+                items: const [
+                  Icon(
+                    Icons.home,
+                    color: Color.fromARGB(255, 255, 255, 255),
+                  ),
+                  Icon(
+                    Icons.search,
+                    color: Color.fromARGB(255, 255, 255, 255),
+                  ),
+                  Icon(
+                    Icons.add,
+                    color: Color.fromARGB(255, 255, 255, 255),
+                  ),
+                  Icon(
+                    Icons.message,
+                    color: Color.fromRGBO(255, 255, 255, 1),
+                  ),
+                  Icon(
+                    Icons.person,
+                    color: Color.fromARGB(255, 255, 255, 255),
+                  ),
+                ]),
+          ),
+        ));
   }
 }
-
-
-
-
 
 Future<bool> showSignOutDialog(BuildContext context) {
   return showDialog<bool>(
@@ -254,4 +232,3 @@ Future<bool> showExitConfirmationDialog(BuildContext context) {
     },
   ).then((value) => value ?? false);
 }
-
