+++ hunch_app/pubspec.yaml
@@ -42,6 +42,8 @@ dependencies:
   http: ^1.1.0
   cloud_firestore: ^4.12.2
   provider: ^6.0.5
+  image_picker: ^1.0.4
+  firebase_storage: ^11.4.1
 
 dev_dependencies:
   flutter_test:
